# Conflicts:
#	EPMS/src/main/resources/application.properties
#	frontend/src/App.tsx
#	frontend/src/components/layout/Sidebar.tsx
#	frontend/src/contexts/AuthContext.tsx
#	frontend/src/pages/employee/EmployeeManagement.tsx

+++ EPMS/src/main/java/com/epms/dto/AppraisalQuestionResponseDto.java
@@ -17,6 +17,6 @@ public class AppraisalQuestionResponseDto {
     private String responseType;
     private Boolean isRequired;
     private Double weight;
-    private Long ratingScaleId;
+    private Integer ratingScaleId;
     private LocalDateTime createdAt;
 }
